+++ src/titiler/core/titiler/core/__init__.py
@@ -1,13 +1,12 @@
 """titiler.core"""
 
-__version__ = "1.1.1"
+__version__ = "2.0.0dev0"
 
 from . import dependencies, errors, factory, routing  # noqa
 from .factory import (  # noqa
     AlgorithmFactory,
     BaseFactory,
     ColorMapFactory,
-    MultiBandTilerFactory,
     MultiBaseTilerFactory,
     TilerFactory,
     TMSFactory,