+++ src/titiler/mosaic/titiler/mosaic/__init__.py
@@ -1,6 +1,6 @@
 """titiler.mosaic"""
 
-__version__ = "1.1.1"
+__version__ = "2.0.0dev0"
 
 from . import errors, factory  # noqa
 from .factory import MosaicTilerFactory  # noqa